+++ pallets/liquidate/src/lib.rs
@@ -35,16 +35,10 @@ pub const KEY_TYPE: KeyTypeId = KeyTypeId(*b"pool");
 pub const LOCK_TIMEOUT_EXPIRATION: u64 = 20000; // in milli-seconds
 pub const LOCK_BLOCK_EXPIRATION: u32 = 10; // in block number
 
-type TotalSumPirce = Price;
-type LiquidationThreshold = u128;
-type DebtAccountBook = (CurrencyId, Balance, Price, TotalSumPirce);
-type CollateralsAccountBook = (
-    CurrencyId,
-    Balance,
-    Price,
-    TotalSumPirce,
-    LiquidationThreshold,
-);
+type TotalValue = u128;
+type LiquidationThreshold = Ratio;
+type DebtAccountBook = (CurrencyId, Balance, Price, TotalValue);
+type CollateralsAccountBook = (CurrencyId, Balance, Price, TotalValue, LiquidationThreshold);
 // the borrower's debt
 type LiquidateToken = CurrencyId;
 // the borrower's collateral
@@ -197,13 +191,10 @@ pub mod module {
                             match <T as module::Config>::PriceFeeder::get_price(currency_id)
                                 .ok_or(Error::<T>::OracleCurrencyPriceNotReady)
                             {
-                                Ok((v, _)) => {
-                                    log::debug!("currency id:{:?}, price: {:?}", currency_id, v);
-                                    v
-                                }
+                                Ok((v, _)) => v,
                                 Err(e) => {
                                     log::error!("error msg: {:?}", e);
-                                    Price::MIN
+                                    MIN_PRICE
                                 }
                             };
                         // 2.1.1 insert debt by currency
@@ -219,11 +210,11 @@ pub mod module {
                                 }
                             };
                         if borrow_currency_amount > 0 {
-                            if currency_price == Price::MIN {
+                            if currency_price == MIN_PRICE {
                                 continue 'outer;
                             }
-                            let borrow_currency_sum_price = match borrow_currency_amount
-                                .checked_mul(currency_price)
+                            let borrow_currency_value = match currency_price
+                                .checked_mul_int(borrow_currency_amount)
                                 .ok_or(Error::<T>::CaculateError)
                             {
                                 Ok(v) => v,
@@ -237,7 +228,7 @@ pub mod module {
                                 *currency_id,
                                 borrow_currency_amount,
                                 currency_price,
-                                borrow_currency_sum_price,
+                                borrow_currency_value,
                             ));
                         }
 
@@ -246,7 +237,7 @@ pub mod module {
                             pallet_loans::AccountCollateral::<T>::get(currency_id, &borrower);
                         //the total amount of borrower's collateral token
                         if collateral_ctoken_amount > 0 {
-                            if currency_price == Price::MIN {
+                            if currency_price == MIN_PRICE {
                                 continue 'outer;
                             }
                             let exchange_rate = pallet_loans::ExchangeRate::<T>::get(currency_id);
@@ -260,10 +251,9 @@ pub mod module {
                                     continue 'outer;
                                 }
                             };
-
                             //the total price of borrower's collateral token
-                            let collateral_currency_sum_price = match collateral_currency_amount
-                                .checked_mul(currency_price)
+                            let collateral_currency_value = match currency_price
+                                .checked_mul_int(collateral_currency_amount)
                                 .ok_or(Error::<T>::CaculateError)
                             {
                                 Ok(v) => v,
@@ -278,7 +268,7 @@ pub mod module {
                                 *currency_id,
                                 collateral_currency_amount,
                                 currency_price,
-                                collateral_currency_sum_price,
+                                collateral_currency_value,
                                 liquidation_threshold,
                             ));
                         }
@@ -291,21 +281,19 @@ pub mod module {
                     // if (collateral_total_value * liquidation_threshold)/(debt_total_value) < 1 ,execute liquidation
                     let mut processing = true;
                     let collateral_liquidation_threshold_value = classify_collaterals.iter().fold(
-                        Price::MIN,
-                        |acc,&(_,_,_,total_sum_price,liquidation_threshold)|
-							// acc + total_sum_price * liquidation_threshold
-							match total_sum_price
-								.checked_mul(liquidation_threshold)
-								.and_then(|r| r.checked_div(RATE_DECIMAL))
-								.and_then(|r| r.checked_add(acc))
-								.ok_or(Error::<T>::CaculateError)
-							{
-								Ok(v) => v,
-								Err(e) => {
-									log::error!("error calculate liquidation threshold: {:?}",e);
-									processing = false;
-									acc
-								}
+                        MIN_PRICE.into_inner(),
+                        |acc,&(_,_,_, total_value,liquidation_threshold)|
+							// acc + total_value * liquidation_threshold
+								match (liquidation_threshold.mul_floor(total_value))
+									.checked_add(acc)
+									.ok_or(Error::<T>::CaculateError)
+								{
+									Ok(v) => v,
+									Err(e) => {
+										log::error!("error calculate liquidation threshold: {:?}",e);
+										processing = false;
+										acc
+									 }
 							},
                     );
                     if !processing {
@@ -313,11 +301,11 @@ pub mod module {
                     }
 
                     let debt_total_value = classify_debts.iter().fold(
-                        Price::MIN,
-                        |acc, &(_,_,_,total_sum_price)|
-							// acc + total_sum_price
+                        MIN_PRICE.into_inner(),
+                        |acc, &(_,_,_, total_value)|
+							// acc + total_value
 							match acc
-								.checked_add(total_sum_price)
+								.checked_add(total_value)
 								.ok_or(Error::<T>::CaculateError)
 							{
 								Ok(v) => v,
@@ -348,11 +336,11 @@ pub mod module {
                         vec![];
 
                     let collateral_total_value = classify_collaterals.iter().fold(
-                        Price::MIN,
-                        |acc,&(_,_,_,total_sum_price,_)|
-							// acc + total_sum_price
+                        MIN_PRICE.into_inner(),
+                        |acc,&(_,_,_, total_value,_)|
+							// acc + total_value
 							match acc
-								.checked_add(total_sum_price)
+								.checked_add(total_value)
 								.ok_or(Error::<T>::CaculateError)
 							{
 								Ok(v) => v,
@@ -366,23 +354,20 @@ pub mod module {
                     if !processing {
                         continue;
                     }
-                    for &(liquidate_token, debt_repay_amount, _, _debt_total_sum_price) in
+                    for &(liquidate_token, debt_repay_amount, _, _debt_total_value) in
                         classify_debts.iter()
                     {
                         let close_factor = pallet_loans::CloseFactor::<T>::get(liquidate_token);
                         //CollateralsAccountBook = (CurrencyId, Balance, Price, TotalSumPirce, LiquidationThreshold);
-                        for &(collateral_token, _, _, single_collateral_total_sum_pirce, _) in
+                        for &(collateral_token, _, _, single_collateral_total_value, _) in
                             classify_collaterals.iter()
                         {
-                            // let repay_amount = (single_collateral_total_sum_pirce / collateral_total_value) * (debt_repay_amount * close_factor);
-                            let m: Price = 100;
+                            // let repay_amount = (single_collateral_total_value / collateral_total_value) * (debt_repay_amount * close_factor);
                             let repay_amount = match (close_factor
-                                * single_collateral_total_sum_pirce)
-                                .checked_mul(m)
-                                .and_then(|r| r.checked_div(collateral_total_value))
-                                .and_then(|r| r.checked_mul(debt_repay_amount))
-                                .and_then(|r| r.checked_div(m))
-                                .ok_or(Error::<T>::CaculateError)
+                                .mul_floor(single_collateral_total_value))
+                            .checked_mul(debt_repay_amount)
+                            .and_then(|r| r.checked_div(collateral_total_value))
+                            .ok_or(Error::<T>::CaculateError)
                             {
                                 Ok(v) => v,
                                 Err(e) => {
@@ -391,7 +376,7 @@ pub mod module {
                                         e
                                     );
                                     processing = false;
-                                    Price::MIN
+                                    MIN_PRICE.into_inner()
                                 }
                             };
                             waiting_for_liquidation_vec.push(WaitingForLiquidation(
@@ -420,6 +405,9 @@ pub mod module {
             // Get signer from ocw
             //TODO get special pool account
             let signer = Signer::<T, <T as module::Config>::AuthorityId>::any_account();
+            if !signer.can_sign() {
+                log::error!("failure: offchain_signed_tx no signer");
+            }
             let result = signer.send_signed_transaction(|_acct|
 				// This is the on-chain function
 				Call::execute_liquidation(waiting_for_liquidation_vec.clone()));